requirements.txt
fastapi>=0.95.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
openai-agents